--- src/stocks_network/multilayer_graph.py
@@ -1,19 +1,0 @@
-import networkx as nx
-from community import community_louvain
-
-# Description
-
-# The MultiLayerGraph class is designed to represent and manipulate multilayer networks,
-# also known as multiplex networks. These networks consist of multiple layers, each 
-# containing a single-layer network (graph), with nodes potentially connected across
-# different layers. The class provides functionalities to create and modify multilayer
-# graphs, add nodes and edges (both intra-layer and inter-layer), and compute various
-# network metrics such as degree sequence, total degree sequence, modularity, and density.
-
-class MultiLayerGraph:
-    def __init__(self, num_layers):
-        self.num_layers = num_layers
-        self.graph = dict()
-
-    def add_layer(self, name, layer):
-        self.graph[name] = layer
